# Relyf.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Relyf API",
    "description": "API for the Relyf upcycling platform - connecting users with creative ways to reuse and repurpose items",
    "version": "v1",
    "contact": {
      "name": "Relyf Team",
      "url": "https://github.com/nanasarf/Relyf"
    }
  },
  "servers": [
    {
      "url": "https://localhost:7139",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Create a new user account with email, password, username, and display name",
        "operationId": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "john@example.com",
                "password": "SecurePass123!",
                "userName": "john_doe",
                "displayName": "John Doe",
                "countryCode": "US"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (validation error)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Username must be between 3 and 20 characters."
              }
            }
          },
          "409": {
            "description": "Email or username already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Username is already taken."
              }
            }
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login with email and password",
        "description": "Authenticate user and receive JWT token",
        "operationId": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "john@example.com",
                "password": "SecurePass123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Email and Password are required."
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Invalid credentials."
              }
            }
          }
        }
      }
    },
...
    "/api/Follow": {
      "post": {
        "tags": ["Follow"],
        "summary": "Follow a user",
        "description": "Create a follow relationship (requires authentication)",
        "operationId": "FollowUser",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Follow created successfully"
          },
          "400": {
            "description": "Cannot follow yourself"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Already following this user"
          }
        }
      }
    },
    "/api/Follow/{followingId}": {
      "delete": {
        "tags": ["Follow"],
        "summary": "Unfollow a user",
        "description": "Remove a follow relationship (requires authentication)",
        "operationId": "UnfollowUser",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "followingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unfollow successful"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Follow relationship not found"
          }
        }
      }
    },
    "/api/Follow/check/{followingId}": {
      "get": {
        "tags": ["Follow"],
        "summary": "Check follow status",
        "description": "Check if current user is following a specific user",
        "operationId": "CheckFollowStatus",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "followingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Follow status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isFollowing": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "userName", "displayName"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "SecurePass123!"
          },
          "userName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "pattern": "^[a-zA-Z0-9_]+$",
            "example": "john_doe"
          },
          "displayName": {
            "type": "string",
            "maxLength": 120,
            "example": "John Doe"
          },
          "countryCode": {
            "type": "string",
            "maxLength": 10,
            "nullable": true,
            "example": "US"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "SecurePass123!"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "example": 123
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "userName": {
            "type": "string",
            "example": "john_doe"
          },
          "displayName": {
            "type": "string",
            "example": "John Doe"
          },
          "token": {
            "type": "string",
            "description": "JWT authentication token",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          }
        }
      },
...
      "FollowRequest": {
        "type": "object",
        "required": ["followingId"],
        "properties": {
          "followingId": {
            "type": "integer",
            "description": "ID of the user to follow"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    }
  },
...
}